Add BigQuery subcriptions to Pubsub module.
Pubsub can now have subscriptions that write directly to BigQuery.
 * https://cloud.google.com/pubsub/docs/bigquery

In the Google Terraform provider, this is configured using an additional block
inside a `google_pubsub_subscription` resource.
 * https://registry.terraform.io/providers/hashicorp/google/latest/docs/resources/pubsub_subscription#nested_bigquery_config

This PR adds a new input variable to the `pubsub` module, to optionally add this
block to some of the subscriptions defined in the module.